Move PYNM_DIR to __init__.py
